cgroups: add mkdir(2), rmdir(2)
Begin support of cgroups by adding mkdir and rmdir. Instead of providing
an inflexible interface to cgroups within the port, support the
primitives so that the logic can be done within erlang.

This functionality duplicates the file module in the standard library.
It is unfortunately necessary because the port requires superuser
privileges to write to the cgroup filesystem.