Support allocating C structs containing pointers
Provide an interace for creating buffers with embedded pointers for
seccomp mode. Seccomp mode uses bpf programs which look like:

    struct sock_fprog {
        unsigned short len;
        struct sock_filter *filter;
    };

The caller can provide generate the structure using a list:

    [
        % 8 instructions in filter
        <<8:2/native-unsigned-integer-unit:8>>,
        % Pad
        <<0:16>>,
        % Generate a binary containing the BPF instructions
        {ptr, bpf_filter()}
    ]

Rough, buggy code to get started testing seccomp mode. In particular,
the contents of the allocated buffers is write only currently. Some of
the prctl() calls are probably in/out.