seccomp: add functions to generate BPF bytecode
ERL_LIBS was pulling in the bpf functions from procket. Copy/paste of
bpf functions from procket into alcove_seccomp. Amazingly, seccomp mode
worked out of the box. Still needs a lot of clean up, as well as a
higher level, simpler, friendlier library to wrap it all up.

The biggest change to follow will be altering the return value of prctl.
Right now, passing in a struct (represented as a list of binaries and tuples)
will return a binary. This isn't very useful because the binary contains
the pointer to the value, not the value itself. With seccomp mode, this
value isn't updated anyway.

prctl/6,7 will change to returning the contents of the pointer. So if a
struct is passed in like:

    [<<"some data">>, {ptr, 6}, {ptr, <<"stuff">>}, <<"foo">>]

The return value will look like:

    [<<"some data">>, {ptr, <<1,2,3,4,5,6>>}, {ptr, <<"stuff">>}, <<"foo">>]