Crash if call/2,3,4,5 reaches the timeout
To ensure consistency, exit if a timeout is set and reached. Otherwise,
late messages may arrive in the queue, messing up subsequent calls:

    1> {ok,P} = alcove_drv:start().
    {ok,<0.45.0>}
    2> catch alcove:call(P, [], getpid, [], 0).
    {'EXIT',timeout}
    3> alcove:version(P).
    2897
    4> flush().
    Shell got {alcove_call,<0.45.0>,[],<<"0.6.1">>}
    ok

Using timeouts might be needed if it is not known whether is still
running in the event loop.

Remove the cast functions, since they are dangerous and can be emulated
by using call/5.