Reset SIGPIPE handler to default
On start up, beam ignores SIGPIPE. Ignored signals are not reset in the
child when exec is called. See sigaction(2):

    A child created via fork(2) inherits a copy of its parent's signal
    dispositions.  During an execve(2), the dispositions of handled  signals
    are reset to the default; the dispositions of ignored signals are left
    unchanged.

This could be considered to be a bug in beam. Usually daemon processes
will ignore SIGPIPE anyway.

Processes that launch a shell should not ignore SIGPIPE, since filter
programs may ignore write failures.

Setting SIGPIPE to the default handler seems to be the most predictable
behaviour for alcove. The downside is that it forces the caller to
manually ignore SIGPIPE on startup:

    SIGPIPE = alcove:define(Drv, 'SIGPIPE'),
    alcove:sigaction(Drv, SIGPIPE, ign).

The alternative is to ignore SIGPIPE and reset SIGPIPE to the defaults
after calling fork() or before calling exec but this would interfere
with any custom signal handlers the user has set up.