Ensure exec returns ok
Remove the check for signals from the code dealing with the control fd.
This ensures a successful call to execvp/3,4 and execve/4,5 that
immediately dies with a signal will return ok, then generate a termsig
event.

It's possible that a non-returning call may be killed by a signal. If
that occurs and the child process exits, there is no way to distinguish
a successful call from a terminated one. In this case, ok will be
returned to the caller, followed by a sigterm event.