Set the control fd's to close on exec
Set the parent's side of the child's fds (stdin, stdout, stderr, fdctl)
to close on exec. This prevents fd's from leaking when the process calls
exec.

Without setting the fd's to close on exec, the child processes spawned
by the exec'ed process would not be terminated until the process exited:

    port -> alcove1 -> alcove2 -> alcove3 -> alcove4

        exec(alcove2, "/bin/cat")

    port -> alcove1 -> /bin/cat -> alcove3 -> alcove4

With this change, the child process exits but becomes a zombie, since
the parent will not have called waitpid to reap the process:

    port -> alcove1 -> /bin/cat -> alcove3 (defunct)