Use the errno for the exit value
Alcove will exit on fatal errors like resource allocation failures. Use
the value of errno as the exit value (which will be sent as a message if
the exit_status option is used).

The reason for the failure is sent to stderr using the err macros and so
will be in format:

    <<"alcove: ", Reason/binary>>