setpriority test: Linux fix when RLIMIT_NICE set
Linux has a resource limit for scheduling priority. If the value is 0,
processes that have increased their nice value (i.e., set it to a
positive integer between 0 and 20) may not decrease it. If the value is
20, the user may decrease the process priority back to 0.

Uniformly test for this resource limit across all OS'es (even though
only linux seems to support it) and set it down to 0 for the test.