alcove_drv: consistently enforce positive PIDs
The port will accept negative PIDs. A PID is usually defined as a signed
32-bit integer.

For fork paths, values will always be positive (as returned by fork;
fork errors will return a tuple rather than -1). If the user passes a
negative number as a component of the fork path, the message will be
discared.