Restrict port usage to owner
Remove the experimental support for multiprocess usage of the port. The
idea was to support one erlang process per Unix process using a single
alcove port, spawning a new process after a fork() or clone, similar to
the way gen_tcp:accept/1 is used.

Multiprocess use of the port will be support in another way: other by
modifying the port protocol to contain the sender's PID or by using an
intermediary process.

Port usage is restricted for calls. After the process has called exec(),
any process can send standard input. stdout is sent to the owner.