Add support for ioctl/5
Control devices using ioctl(2).

ioctl's using an integer for argp has not been tested.

ioctl/5 requires an fd previously opened using open/5. The behaviour
is inconsistent from setns/4 which takes a path and then opens an fd on
behalf of the caller. setns/4 predates the introduction of open/5.

Providing the argument as a path is simpler for the caller. Providing an
fd is more flexible: for example, a process can open the fd and call
clone() to create a child in a namespace. The child can drop privs and
call ioctl() on the inherited fd.

So either:
* the setns interface should be changed to take fd's
* supporting both fd's and paths in setns and ioctl at the expense of
  duplicating code