ioctl: fix compiler warning
The compiler was warning about the branches of the ternary conditional
returning different types (int and pointer to char). Duplicate the
ioctl's using a switch statement instead.

Note the third argument (argp) to ioctl(2) is not specified to avoid
compiler type warnings.

There is another issue to be fixed: the second argument (variously
referred to as request or command) is an unsigned long on FreeBSD and
Mac OS X. The code will need to be appropirately ifdef'ed for those
platforms.